P>Q
- adds an additional state to price (below, above, above Q)
- updates cases to reflect (+ removed null cases)
- updated sop cases
- updated tests to check P>Q cases

+++ protocol/contracts/libraries/LibGauge.sol
@@ -151,7 +151,7 @@ library LibGauge {
         if(depositedBdv > 0){
             LibWhitelist.updateStalkPerBdvPerSeasonForToken(
                 token,
-                uint24(newGrownStalk.div(depositedBdv))
+                uint32(newGrownStalk.div(depositedBdv))
             );
         }
     }
